+++ inst/doc/simTool.R
@@ -196,6 +196,61 @@ someFunc <- function() {
 someFunc()
 
 ## -------------------------------------------------------------------------------------------------
+dg <- expand_tibble(fun = c("rnorm"), mean = c(1,1000), sd = c(1,10), n = c(10L, 100L))
+pg <- expand_tibble(proc = "quantile", probs = 0.975)
+post_ana <- function(q_est, .truth){
+  tibble::tibble(bias = q_est - stats::qnorm(0.975, mean = .truth$mean, sd = .truth$sd))
+}
+eval_tibbles(dg, pg, replications = 10^3, discard_generated_data = TRUE,
+                   ncpus = 2,
+                   post_analyze = post_ana,
+                   summary_fun = list(mean = mean))
+
+## -------------------------------------------------------------------------------------------------
+dg <- dplyr::bind_rows(
+  expand_tibble(fun = c("rnorm"), mean = 0, n = c(10L, 100L), .truth = qnorm(0.975)),
+  expand_tibble(fun = c("rexp"), rate = 1, n = c(10L, 100L), .truth = qexp(0.975, rate = 1)),
+  expand_tibble(fun = c("runif"), max = 2, n = c(10L, 100L), .truth = qunif(0.975, max = 2))
+)
+pg <- expand_tibble(proc = "quantile", probs = 0.975)
+post_ana <- function(q_est, .truth){
+  ret <- q_est - .truth
+  names(ret) <- "bias"
+  ret
+}
+eval_tibbles(dg, pg, replications = 10^3, discard_generated_data = TRUE,
+                   ncpus = 2,
+                   post_analyze = post_ana,
+                   summary_fun = list(mean = mean))
+
+## -------------------------------------------------------------------------------------------------
+dg <- dplyr::bind_rows(
+  expand_tibble(fun = c("rnorm"), mean = 0, n = c(10L, 1000L), 
+                .truth = list(function(prob) qnorm(prob, mean = 0))),
+  expand_tibble(fun = c("rexp"), rate = 1, n = c(10L, 1000L),
+                .truth = list(function(prob) qexp(prob, rate = 1))),
+  expand_tibble(fun = c("runif"), max = 2, n = c(10L, 1000L),
+                .truth = list(function(prob) qunif(prob, max = 2)))
+)
+bias_quantile <- function(x, prob, .truth) {
+  est <- quantile(x, probs = prob)
+  ret <- est - .truth[[1]](prob)
+  names(ret) <- "bias"
+  ret
+}
+pg <- expand_tibble(proc = "bias_quantile", prob = c(0.9, 0.975))
+eval_tibbles(dg, pg, replications = 10^3, discard_generated_data = TRUE,
+                   ncpus = 1,
+                   summary_fun = list(mean = mean))
+
+## ---- eval = FALSE--------------------------------------------------------------------------------
+#  6.  for g in  {g_1, ..., g_k}
+#  7.      for r in 1:replications (optionally in a parallel manner)
+#  8.          data = g()
+#  9.          for f in  {f_1, \ldots, f_L}
+#  10.             append f(data) to the result object (optionally apply a post-analyze-function)
+
+## -------------------------------------------------------------------------------------------------
 EVAL <- FALSE
 if (Sys.getenv("NOT_CRAN") == "true") {
   EVAL <- TRUE
@@ -233,14 +288,14 @@ bootstrap_ci <- function(x, conf.level, R = 999) {
 }
 
 ## ---- eval = EVAL---------------------------------------------------------------------------------
-post_analyze <- function(o) {
+post_analyze <- function(o, .truth) {
   if (class(o) == "htest") {
   #post-process the object returned by t.test
     ci <- o$conf.int
     return(tibble::tibble(
       type = "t.test",
       aspect = c("covered", "ci_length"),
-      value = c(ci[1] <= 3L && 3L <= ci[2], ci[2] - ci[1])
+      value = c(ci[1] <= .truth && .truth <= ci[2], ci[2] - ci[1])
     ))
   } else if (class(o) == "bootci") {
   #post-process the object returned by boot.ci
@@ -255,7 +310,7 @@ post_analyze <- function(o) {
       tibble::tibble(
         type = type, 
         aspect = "covered", 
-        value = as.integer(lower <= 3L & 3L <= upper)),
+        value = as.integer(lower <= .truth & .truth <= upper)),
       tibble::tibble(
         type = type, 
         aspect = "ci_length", 
@@ -266,9 +321,9 @@ post_analyze <- function(o) {
 
 ## ---- eval = EVAL---------------------------------------------------------------------------------
 dg <- dplyr::bind_rows(
-  simTool::expand_tibble(fun = "rnorm", n = 10L, mean = 3, sd = sqrt(3)),
-  simTool::expand_tibble(fun = "runif", n = 10L, max = 6),
-  simTool::expand_tibble(fun = "rexp", n = 10L, rate = 1 / 3)
+  simTool::expand_tibble(fun = "rnorm", n = 10L, mean = 0, sd = sqrt(3), .truth = 0),
+  simTool::expand_tibble(fun = "runif", n = 10L, max = 6, .truth = 3),
+  simTool::expand_tibble(fun = "rexp", n = 10L, rate = 1 / sqrt(3), .truth = sqrt(3))
 )
 dg
 
@@ -296,4 +351,61 @@ et$simulation %>%
   geom_label(position = position_dodge(0.9), size = 3) +
   theme(legend.position = "bottom") + 
   facet_grid(aspect ~ conf.level, scales = "free")
+
+## ---- eval = EVAL---------------------------------------------------------------------------------
+t_test = function(x, conf.level){
+  tt <- t.test(x, conf.level = conf.level)
+  
+  # unify return
+  tibble::tibble(type = "t.test", lower = tt$conf.int[1], upper = tt$conf.int[2])
+}
+
+bootstrap_ci <- function(x, conf.level, R = 999) {
+  b <- boot::boot(x, function(d, i) {
+    n <- length(i)
+    c(
+      mean = mean(d[i]),
+      variance = (n - 1) * var(d[i]) / n^2
+    )
+  }, R = R)
+  ci <- boot::boot.ci(b, conf = conf.level, type = "all")
+  method = c("normal", "basic", "student", "percent", "bca")
+  ret = ci[method]
+  lower = unlist(purrr::map(ret, ~dplyr::nth(.x, -2)))
+  upper = sapply(ret, dplyr::last)
+  type = paste("boot", method, sep = "_")
+  
+  # unify return
+  tibble::tibble(type = type, lower = lower, upper = upper)
+}
+
+dg <- dplyr::bind_rows(
+  simTool::expand_tibble(fun = "rnorm", n = 10L, mean = 0, sd = sqrt(3), .truth = 0),
+  simTool::expand_tibble(fun = "runif", n = 10L, max = 6, .truth = 3),
+  simTool::expand_tibble(fun = "rexp", n = 10L, rate = 1 / sqrt(3), .truth = sqrt(3))
+)
+
+pg <- simTool::expand_tibble(
+  proc = c("t_test","bootstrap_ci"),
+  conf.level = c(0.9, 0.95)
+) %>% 
+  mutate(R = ifelse(proc == "bootstrap_ci", 100, NA))
+
+et <- eval_tibbles(dg, pg,
+  replications = 10^2, ncpus = 2
+)
+et
+
+## ---- eval = EVAL---------------------------------------------------------------------------------
+grps <- et$simulation %>% 
+  select(-replications) %>% 
+  select(fun:type) %>% 
+  names
+
+et$simulation %>% 
+  mutate(covered = lower <= .truth & .truth <= upper,
+         ci_length = upper - lower) %>% 
+  group_by_(.dots = grps) %>% 
+  summarise(coverage = mean(covered),
+            ci_length = mean(ci_length))