+++ NEWS.md
@@ -1,3 +1,13 @@
+Version 1.1.1
+=========================
+
+Misc: 
+
+* .truth-functionality added, i.e. the parameters of the data generation (or alternatively a column of the data generating matrix with the name .truth) is passed to the the data analyzing functions, see the vignette for details
+* Unnesting of the simulation results improved
+
+
+
 Version 1.1.0
 =========================